// fronted/src/generated/models/AnswerDetailVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AnswerDetailVO = {
    questionId?: string;
    subject?: string;
    question?: string;
    answerA?: string;
    answerB?: string;
    answerC?: string;
    answerD?: string;
    rightAnswer?: string;
    analysis?: string;
    score?: number;
    section?: string;
    level?: string;
    answer?: string;
    type?: string;
};