// fronted/src/generated/models/ExamManage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExamManage = {
    examCode?: number;
    description?: string;
    source?: string;
    paperId?: number;
    examDate?: string;
    totalTime?: number;
    grade?: string;
    term?: string;
    major?: string;
    institute?: string;
    totalScore?: number;
    type?: string;
    tips?: string;
};