// fronted/src/generated/models/FillQuestion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FillQuestion = {
    questionId?: number;
    subject?: string;
    question?: string;
    answer?: string;
    score?: number;
    level?: string;
    section?: string;
    /**
     * 题目解析
     */
    analysis?: string;
};