// fronted/src/generated/models/JudgeQuestion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JudgeQuestion = {
    questionId?: number;
    subject?: string;
    question?: string;
    answer?: string;
    level?: string;
    section?: string;
    score?: number;
    /**
     * 题目解析
     */
    analysis?: string;
};