// fronted/src/generated/models/MultiQuestion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MultiQuestion = {
    questionId?: number;
    subject?: string;
    section?: string;
    answerA?: string;
    answerB?: string;
    answerC?: string;
    answerD?: string;
    question?: string;
    level?: string;
    rightAnswer?: string;
    /**
     * 题目解析
     */
    analysis?: string;
    score?: number;
};