// fronted/src/generated/models/Score.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Score = {
    examCode?: number;
    studentId?: number;
    subject?: string;
    ptScore?: number;
    etScore?: number;
    score?: number;
    scoreId?: number;
    answerDate?: string;
};