// fronted/src/generated/models/Student.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Student = {
    studentId?: number;
    studentName?: string;
    grade?: string;
    major?: string;
    clazz?: string;
    institute?: string;
    tel?: string;
    email?: string;
    pwd?: string;
    cardId?: string;
    sex?: string;
    role?: string;
};