// fronted/src/generated/models/Teacher.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Teacher = {
    teacherId?: number;
    teacherName?: string;
    institute?: string;
    sex?: string;
    tel?: string;
    email?: string;
    pwd?: string;
    cardId?: string;
    type?: string;
    role?: string;
};